feat : Get 요청에 대해 로그인 여부를 확인하고 알맞은 핸들러로 분기시키는 기능을 하는 객체 구현
Get 요청이 들어오면 로그인 여부를 체크 후, 알맞은 핸들러로 분기해 준다
